[kernel][vm] Retain AwaitingCleanLength if possible
When a zero interval is split in the middle or clipped from the end, its
AwaitingCleanLength still holds. Similarly when the interval is split at
the start, its AwaitingCleanLength (minus a page) should be carried over
to the new start on the right. This ensures the AwaitingCleanLength is
retained across all kinds of splits. The same holds for when an interval
is clipped at the start, or when slots are populated starting at the
interval start.

Conversely, when unused interval slots are returned via
ReturnIntervalSlot, their AwaitingCleanLength should be merged back into
the resultant interval.

The only case we cannot retain AwaitingCleanLength is when an interval
is extended / merged on the left, because the AwaitingCleanLength only
describes a range that begins at the interval start.

Test: vmpl_unittest
Bug: 122842
Run-All-Tests: true
Change-Id: Idc7bd6b139859fdea0284eb6a1d0c84672858518
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837216
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>