[ffx] add daemon.autostart[=true] flag
If daemon.autostart is disabled, then the daemon will not magically
start -- instead the message "FFX Daemon was told not to autostart
and no existing Daemon instance was found" will be produced.

This gives EngProd the ability to control the lifecycle of the
daemon directly.

Fixed: 120758
Change-Id: Id9524af29451bcd162405100c5dedf9d1efc1402
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837996
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>