[starnix] Don't log a warning on vfork.
Starnix doesn't yet implement `clone` with `CLONE_VM` without `CLONE_THREAD`. As a workaround, it behaves like a fork, and logs a warning when that happens. This works for most use cases, and the warning helps identify new possible problematic uses.

`vfork` uses the same code path, but it is OK to implement `vfork` like `fork`, so there's no need to warn the user about it.

Bug: 114813
Change-Id: I89a55902ef03b94f659500b6da18d276ff4e60c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837876
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>