[kernel][scheduler] Remove GetTargetPreemptionTime
Remove the GetTargetPreemptionTime method from the Scheduler.  It is currently
unused.  If similar functionality is needed at some point in the future, the
method can be re-introduced in a way which is compatible with the
thread-lock-free future which is fast approaching us.

Change-Id: I4a5bb17b456a3ded8de0f759e98507092d495f3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839036
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>