[driver_runtime] Allow multiple dispatchers per thread pool.
The coordinator checks whether a thread pool with the same
scheduler role already exists.

A thread wakeup handler has also been added to the thread
pool's async loop, which will check whether the scheduler
role has been set for the current thread. This is as we
can't post a task to a specific thread in the pool.

Bug: 124644,85946

Change-Id: Id5f35c3fbe94682aaf196799e8f1d6686d47bc8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837857
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>