[fx] Disable report-bug inside fuchsia.git
This change modifies `report-bug` to warn that it is no longer usable
and exit. Components are no longer specified in OWNERS files after the
submission of
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/834141

This file is an exceptional case of a haunted graveyard and is
essentially impossible to modify by any standard software engineer.
Because of this, it needs to be considered legacy code.

Analytics shows it is in single digit usage of the past month.

Bug: b/278592638
Change-Id: Ia7cd383221c6c737c460fcea2d5ac42e6cc60671
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838056
Commit-Queue: Chris Lewis <cflewis@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>