[starnix] Add a test for mprotect failure.
This CL adds a syscall test that makes sure starnix "doesn't get
confused" if `mprotect` fails. The idea is that `mprotect` could fail on
any part of the provided address range, and it is not defined what
should happen to the other parts of the range.

The test make sure that only two options can happen: either the mprotect
change succeeded or it failed. In both cases, it also tries to access
the memory regions to make sure that the actual mapping rights
corresponds with the reported ones.

Test: fx test syscalls_test --test-filter "*MprotectFailureIsConsistent*"
Bug: 123205
Change-Id: I8bc40cb8e4960129b40e31bda9d9a8f9f1479595
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824390
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Marco Vanotti <mvanotti@google.com>