[ffx] Only do extreme build configs for integrated ffx
When building ffx for separate subtools the binary is much smaller and
we can relax a bit in terms of our optimization choices. This changes
the ffx template to allow for this and uses the separate subtools flag
to choose the value of that flag.

Change-Id: I50c75c354312f1b1909f197225730e26cff66c4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/836156
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>