[check-licenses] Updates to file package.
In the future, license analysis will run in multiple threads,
so each package needs to be concurrency-safe.

This CL un-exports all public fields in the file package, opting
instead to use getters and setters everywhere. This is necessary
to prevent race conditions.

It also introduces custom JSON marshalling logic, since fields need
to be exported in order to be read by the default json marshaller.

Finally, it generates the SPDX IDs for files and filedata objects
by using a hash of the data. This should result in more deterministic
output files.

Change-Id: I7e550b1db7ddb394ceb170815c84b317d4716f62
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838096
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Yaar Schnitman <yaar@google.com>