[fshost] Handle zxcrypt wrapping inside ZxcryptDevice.
This removes the need for plumbing of inside_zxcrypt and special casing.

There are still a few asymmetries here that are a bit ugly IMHO, but
I'll leave those as an exercise for the future. For instance, it
would be nice if SetPartitionMaxBytes was a Volume protocol call rather
than VolumeManager as this would allow zxcrypt devices to transparently
account for their own overheads and further simplify code.

Change-Id: I0c8c73204d85a413302a982443c976af2b8d69ee
Bug: 123427
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837237
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Aaron Drew <ripper@google.com>