[kernel][vm] Use LookupCursor
Replaces usages of LookupPagesLocked with the stateful LookupCursor.

This impacts a lot of microbenchmarks, both positively and negatively
can be broken down into a few groups.

1. 1-2% drop in some stream benchmarks. Testing shows that this
reduction largely stems from a change in inlining, with the massive
LookupPagesLocked now having it's logic split across multiple functions,
and the compiler not neccessarily always choseing to inline. Some
__ALWAYS_INLINE tags have already been placed close this overhead, and
the remaining slowdown can largely be fixed with further force
inlining, however it's not clear we should be agressively inlining just
based on microbenchmarks from one platform. For now a 1-2% drop is
acceptable.

2. 0-1% drop in many VMO clone benchmarks. Again, this is due to
inlining changes, with the difference to the stream case being that
the overhead is even less, and the importance of the performance of these
cases also being less.

3. Various mapping related benchmarks improve by ~5%, with
"Heap/create_1mb_vmo" by 20+%.

Overall the improved readability of the new cursor, combined with the
performance improvements, make this worthwhile. The performance losses
can be regained by forcing inlining if deemed neccessary.

Change-Id: Icfbe96a86c58f7bb2f0088f893325a0ce32b9b72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797766
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>