[starnix] Request backtrace in CoreDump crashes in restricted mode
When the restricted mode executor is about to exit a task due to the
taks state being set to a CoreDump exit code, solicit a backtrace from
the system by generating a "backtrace request" in the form of a software
breakpoint exception with a magic value in rax / x0. This also (re)emits
CFI directives to help unwinders see into the Linux state at the point
of exit. This is particularly useful for Linux abort() implementations
that raise the SIGABRT signal on themselves.

Crashsvc's inspector library cannot correctly unwind this stack yet, but
it should be able to soon.

Bug: 124427
Change-Id: Icb2cd998167636f3ce46fb7fdcf241227e267388
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838478
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>