[rbe] Auto-relaunch self with reproxy
Include a unique build-id (e.g. the reproxy log dir) in download stub.

The unique build-id will inform the RBE wrapper whether the stub
came from the same build session as the current one, or a previous one.
Within the same build session, it is guaranteed that the RBE CAS will
already hold the referenced object.  Across different build sessions
(like with incremental builds), there may be a way to inform
reproxy to use an existing file metadata if the stub is present.

Download stubs are written to the same location as the original output,
so local consumers will need to be informed of this to auto-fetch
inputs.

For the build-id, we arbitrarily chose to use the reproxy tempdir,
which is setup by fuchsia-reproxy-wrap.sh.

Since build-id is now required for remote execution, we also make
the reproxy wrap more automatic, without requiring `--auto-reproxy`.
This should make repro-ing individual actions easier, without having to
remember --auto-reproxy, which can be considered deprecated now.
Users will see a notices that the option is no longer needed.

Both the C++ and Rust wrappers auto-relaunch themselves.

Bug: 123178
Fixed: 124190

Change-Id: Ifdb44d2ce09b0a878e582dfd73590afe2fc006a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839762
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: David Turner <digit@google.com>