[wlan] Modify softmac beaconing APIs
Removes ConfigureBeaconing() to disallow updates to beaconing once it is
started. Right now the API surface only provides an EnableBeaconing()
and a DisableBeaconing() to keep it simple. When there is a need to
update beaconing, the suggestion is to create a new method named
UpdateBeaconing() or such, and pass it only the set of parameters that
we would like to allow in an update method.

Bug: 87050

Change-Id: Icb5383d69f41a7bdca621e78a6eb8c1124a5ffb9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831321
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Fuchsia-Auto-Submit: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Commit-Queue: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>