[radar][proxy] Use platform config to enable proxying on eng/userdebug
proxy_radar_burst_reader tells radar-proxy whether to act as a proxy for
radar data or to allow clients to bypass it and connect directly to the
driver. Connecting directly is better for performance, but prevents
burst injection testing. Use platform configuration to enable proxying
on eng and userdebug builds, but disable it on user builds.

Bug: 123714
Test: Radar features work on Nelson
Test: fx test //src/devices/radar/test/radar-integration-test
Change-Id: I982122d5a36b5857a027735cd98e3acbe70b2c59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/839576
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>