[dsi-dw] Remove logging macros.
This CL removes the logging macros in dsi-dw, because the C++ style
guide mandates that macros are only used when absolutely necessary.

The logging macros are replaced with direct zxlogf() calls. The extra
functionality in the macros is no longer necessary, because zxlogf()
includes the file name and line number in logging output. While at it,
this CL also does minor logging improvements, such as replacing numeric
zx_status_t values with string representations, and clarifying the
wording of the logging messages.

Test: fx test dsi-dw-test
Test: Booted core.vim3 with Khadas TS-050
Test: fx shell display-tool squares
Change-Id: Ic9a9ec7e1f355b8bcff63adbb482169be7f69797
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/840837
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>