[bt][hfp] Restart Codec Connection Setup on fail
From HFP 1.8:

If an (e)SCO link cannot be established according to the parameters
required for the selected codec (e.g., basebands negotiation fails for a
link with re-transmission although a wide band codec has been selected),
the Codec Connection establishment procedure shall be re-started by the
AG with the purpose of selecting a codec that can be used. The mandatory
narrow band Codec (CVSD) shall be used before the AG gives up trying to
establish a Codec connection.

This means we need to re-select the CVSD codec after failing to connect
for a higher quality codec, and before initiating the CVSD connection.

Fixes: 123933
Test: added test; fx test bt-hfp-audio-gateway-tests

Change-Id: Iedf4246a57e2314825b95f3d6f46bf0764463e76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824978
Reviewed-by: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>