[ffx][emu] Reading port mapping from qmp
This removes the use of the "port-picker" library to
attempt to pre-allocate TCP ports to use with user mode
networking.

Now, the ports are passed in unassigned and assign internally to QEMU
and then read via the QEMU management console socket.

The effect will be eliminating chances of port in use collisions,
especially in CI/CQ environments.

Bug: 121732
Bug: 103540

Change-Id: I11b38ab0770cdaa1333470fc675ac06f4fce1738
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811189
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>