[amlogic-display] Use 16x16 SPLIT_BLOCK formats for AFBC by default.
Upstream amlogic-display driver requires SPLIT_BLOCK modifier bit
set for AFBC images with 16x16 block, or 4k AFBC images with 32x8
block. If SPLIT_BLOCK is not specified, for large images (e.g.
width = 1920), there'll be blocks / lines of pixels displayed as
random colors.

Besides for better performance, we also added YUV and TILED_HEADER
modifiers; these format modifiers are not used in upstream amlogic
driver but we have verified it's working for vulkan applications.

A side effect of this change is that it breaks display-test, as
the encoding logic of display-test is not compatible with the
TILED_HEADER and YUV modifier we added to the display-driver.
fxbug.dev/125730 tracks the effort to re-enable the test for
amlogic-display.

Test: vkcube-on-fb on astro, no regression (afbc enabled)
Test: vkcube-on-fb on vim3, no regression (afbc disabled)
Test: vkcube-on-fb on vim3, after patching fxrev.dev/838618,
      afbc can be enabled
Bug: 125730
Bug: 124949
Change-Id: Iadec3075fc71ed7c1e12505c65ac94d464e12705
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838617
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: John Bauman <jbauman@google.com>