[build][infra] Install all ffx tools for botanist
This makes it so that when running infra tests, we install ffx and all
the separately compiled subtools that might be used by infra tests. It
also sets the go ffxutil library to look for subtools in the place where
they'll be.

Fixed: 125363

Change-Id: I1779c9c8cb2be28a8f0858720fbf98357b16f9fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/826298
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>