[starnix] Return the restricted reason code to rust
Update the x64 and arm64 trampolines to accept a caller provided
pointer. When returning from restricted mode, the kernel supplied
"reason code" will be written to this pointer so that the restricted
executor can take the appropriate action.

The kernel will always set this to ZX_RESTRICTED_REASON_SYSCALL but in
the future it will begin to also generate ZX_RESTRICTED_REASON_EXCEPTION
exits.

Bug: 121512
Change-Id: I35a11f249443f0c785107b7b33645c3ba4c85a3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/840776
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>