[fdf][ti-tca6408a] Forward metadata
Forward the GPIO pins metadata when creating the ti-tca6408a device so
that the gpio driver can access it when the driver binds to the device.

Originally in dfv1, the gpio driver was still able to get the metadata
because in dfv1 the driver manager searches the entire device tree for
the metadata. In dfv2 however, the metadata is searched for only in the
parent device causing the search to fail.

Bug: 125848
Change-Id: Ia9c887c4fb3f0fad7989793bb9208ad699388afc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841556
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>