[kernel] Update 'error code' in r-mode tests
These tests were returning an 'error code', which is somewhat ambiguous
in the context of these tests. Instead we rename this to
'reason code' to match the terminology used in the restricted mode
documentation, and also update the vector code to populate this with the
exception code provided by zircon.

Bug: 121512
Change-Id: If92566dfac9666a567218431a90f7d685f723880
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830782
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>