[audio][tests] Add root driver
Add a root driver that binds to the test root and creates a node
for each child driver.

This is preferred because the driver framework is about to stop
allowing multiple drivers to bind to the test root.

Bug: 125832
Change-Id: I57230ff454335b649c4c235f1ec87d1641310fcb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842396
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>