[build] Always use -msmall-data-limit=0 for RISC-V
This will become the default in the compiler (or become
unnecessary). Use of .sdata/.sbss is technically independent from
linker gp relaxation and theoretically still useful on its own.
However, the compiler moves .rodata and RELRO data to .sdata,
which is problematic for some Fuchsia code. So the compiler will
be changing under --target=riscv64-fuchsia either to default to
-msmall-data-limit=0 in all cases or else instead never to move
.rodata/RELRO objects into .sdata. Until then, explicitly pass
-msmall-data-limit=0 in cases where it's not already the default.

Bug: 125687
Change-Id: Id9742589824c2691ab86aa1393fa0b3424ef020e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842539
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>