[build] Factor hermetic_code_blob() out of code_patching_hermetic_alternative()
This refactors the machinery used for hermetic leaf alternatives
in the kernel code-patching support into a new general-purpose
hermetic_code_blob() template. This can be used for all cases of
hermetic code blob linking, not just the code-patching case. It
optionally supports a code blob with .rodata as well as .text
sections, which is not allowed in the code-patching case. It also
has machinery to make it convenient to share a local API header
between the hermetic code and other code that uses the blob.

Bug: 125717
Change-Id: Idd6d799b2fd4a0640c4a426b3ea0f6778b519cd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841116
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>