[mini-process] Revamp build of subprocess code
This updates the build of the mini-process library without much
change to its actual code. Heretofore, the code blob to be copied
into the subprocess (source in subprocess.cc) was built as a
normal function in the library code, with some tweaks and luck
that kept a self-contained contiguous code blob that could be
copied in practice, but no guarantee the compiler would keep
producing that. For riscv64, the compiler generates .rodata
constants and outcalls to things like memset or memcpy, and such
was always possible on any machine.

The new approach builds the subprocess code on its own and links
it into a flat image separate from the library code. This is then
embedded as a raw blob in the .rodata of the library, where it
can be copied from easily (but cannot be accidentally executed in
the controlling program that links in the mini-process library,
which was possible before). The new special-case link permits use
of non-code sections like .rodata, and also provides standard C
string functions that can be linked in as needed.

Bug: 125717
Change-Id: I08cfc5f574934f4851071b7a13e66848b2577507
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841117
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>