[blobfs] skip blobs that aren't written in health check
If a blob has size zero in the health check, skip it. This is either
because it's the null blob, or because it's not in the kReadable state.
Either way it's not super useful to health-check it.

This hardens the health check so it can be run while blobs are being
written.

Bug: 124968
Test: fx test blobfs-integration.cm
Change-Id: Ie35ea533771f42dce91170c49bc3dbdbc9b2b510
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842856
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>