[netstack3] Select over watcher stream and socket readable
A stream socket's send task uses a `NeedsDataWatcher` to subscribe to a
stream of wakeups that tell it it may be unblocked and should attempt to
do some work: namely, reading from the zircon socket and calling into
Core to send out that data. Currently, when the send task is woken up by
the `NeedsDataNotifier`, it initiates a blocking wait for the zircon
socket to become readable. This means that if the watcher stream
receives another notification while the send task is still waiting for
the socket to become readable, it cannot observe that notification.

One example of this scenario is when a stream socket is closed. When
this happens, Core drops the notifier, which wakes the waker for the
`NeedsDataWatcher`. This is so whoever is holding the watcher stream can
see that it has ended and clean up any resources it's using. However, if
the send task is currently blocked on the aforementioned wait for the
zircon socket to become readable when the stream socket is closed, then
it's effectively stuck in limbo forever: the zircon socket will never
become readable, and Core cannot signal to the send task that it should
stop waiting. Further, because the send task holds a strong reference to
the zircon socket, the socket will never be dropped, causing a leak.

The send task should instead select over both the ZX_SOCKET_READABLE
signal being asserted and the watcher receiving another notification.
This allows the send task to be interrupted when the watcher stream ends
and correctly drop the zircon socket.

Change-Id: Ie41cf89fc89940aa91c68d1af6ca144fff2c7257
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842538
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zeling Feng <zeling@google.com>