[log][rust] Allow the initialization library to work on the host
This change also implements an Event Formatter which is basically the
Compact formatter with support for tags. The goal is to support well the
fuchsia crate which allows clients to set tags and are currently ignored
when using the default compact formatter.

Fixed: 110026

Change-Id: I62269622950ef88f43271d9dc8fe46307b586e43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/842278
Reviewed-by: Brian Bosak <bbosak@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>